mon: Monitor: init clog to output to syslog (or not)
The LogClient has changed in a way that it is now feasible having more
than one LogClient per daemon/client.  As such, the LogClient can no
longer rely on daemon/client-specific boolean values to know whether it
should log stuff to syslog or not.  Instead of breaking API
compatibility, we now delegate to the user of LogClient the
responsibility of informing LogClient on whether it should or should not
log to syslog.  Therefore we need to have the monitor setting 'clog' to
whatever value is in conf->clog_to_syslog.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>