mds/RecoveryQueue: add method to prioritize a file recovery; fix logging
Add a prioritize() method to make file recovery start immediately for the
given inode.  Note that this doesn't respect the max recovery limit: if
someone stats it, they are blocking, and we start the recovery immediately.

Also fix up the dout logging a bit so that everything is prefixed
consistently.

Signed-off-by: Sage Weil <sage@redhat.com>